Split up collision functions and classes in different files
