+++ src/physics/CMakeLists.txt
@@ -2,4 +2,10 @@ set(PHYSICS_SOURCES
         src/physics/physics_system.h
         src/physics/physics_system.cpp
         src/physics/collision.h
-        src/physics/collision.cpp)
+        src/physics/collision.cpp
+        src/physics/solver.h
+        src/physics/solver.cpp
+        src/physics/simplex.h
+        src/physics/algorithm.h
+        src/physics/algorithm.cpp
+        )
