+++ src/physics/solver.h
@@ -1,0 +1,14 @@
+#pragma once
+#include "collision.h"
+
+namespace physics {
+    class Solver {
+    public:
+        virtual void solve(std::vector<Collision> &collisions, float dt) = 0;
+    };
+
+    class PositionSolver : public Solver {
+    public:
+        void solve(std::vector<Collision> &collisions, float dt) override;
+    };
+}
